Create Code
This is the code of the Tabulation Algorithm. It includes a test case at the end, and for it to work, the election results (data frame with all votes cast) needs to be in the same folder as the algorithm. Moreover, it will need to be a csv, and you will need to update the "file_path" to match the name of the csv.